+++ src/v1/client/tasks.rs
@@ -3,7 +3,7 @@
 /// An argument that affects which fields are returned on certain task-related
 /// endpoints.
 
-#[derive(Debug, Default, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[cfg_attr(feature = "serde", serde(rename_all = "UPPERCASE"))]
 pub enum View {