+++ src/v1/types/responses.rs
@@ -6,7 +6,7 @@ pub mod task;
 pub use service::ServiceInfo;
 
 /// A response from `POST /tasks`.
-#[derive(Debug, Default, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct CreateTask {
     /// The ID of the created task.
@@ -14,7 +14,7 @@ pub struct CreateTask {
 }
 
 /// The response from `GET /tasks`.
-#[derive(Debug, Default, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ListTasks<Task> {
     /// The tasks in this page of results.