+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Promotes the `v1::types::responses::service` module to public.
 - Gates serialization/deserialization behind the `serde` feature.
+- Makes most structs `Clone`.
 
 ## 0.2.0 - 08-08-2024